isolate coremod conditional logic
