+++ src/main/java/com/theishiopian/parrying/CoreMod/Hooks/PlayerHooks.java
@@ -1,4 +1,4 @@
-package com.theishiopian.parrying.Mixin;
+package com.theishiopian.parrying.CoreMod.Hooks;
 
 import com.theishiopian.parrying.Config.Config;
 import com.theishiopian.parrying.Items.ScopedCrossbow;
@@ -7,21 +7,15 @@ import com.theishiopian.parrying.Registration.ModItems;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(Player.class)
-public class PlayerMixin
+import java.util.Optional;
+
+public class PlayerHooks
 {
-    @Inject(at = @At("HEAD"), method = "getCurrentItemAttackStrengthDelay", cancellable = true)
-    private void InjectIntoGetCurrentItemAttackStrengthDelay(CallbackInfoReturnable<Float> cir)
+    public static Optional<Float> ModifyAttackStrength(Player player)
     {
         if(Config.dualWieldEnabled.get())
         {
-            Player player = ((Player)(Object)this);
-
             if(DualWieldingMechanic.IsDualWielding(player))
             {
                 float mainSpeed = (float) player.getMainHandItem().
@@ -35,20 +29,22 @@ public class PlayerMixin
                 float speedMod = (mainSpeed + offSpeed) / 2;//average speed
                 float diff = Math.abs(mainSpeed - offSpeed);//penalty for difference in speeds
 
-                cir.setReturnValue((float)((1.0D / (Attributes.ATTACK_SPEED.getDefaultValue() + speedMod + 0.2f - diff)) * 20.0D));
+                return Optional.of((float) ((1.0D / (Attributes.ATTACK_SPEED.getDefaultValue() + speedMod + 0.2f - diff)) * 20.0D));
             }
         }
+
+        return Optional.empty();
     }
 
-    @Inject(at = @At("HEAD"), method = "isScoping", cancellable = true)
-    private void InjectIntoIsScoping(CallbackInfoReturnable<Boolean> cir)
+    public static Optional<Boolean> ModifyScopingStatus(Player player)
     {
-        Player player = ((Player) (Object) this);
         boolean isUsingScopedWeapon = player.isUsingItem() && player.getUseItem().is(ModItems.SCOPED_CROSSBOW.get());//todo interface
         if(isUsingScopedWeapon)
         {
             boolean hasProjectile = ScopedCrossbow.isCharged(player.getUseItem());
-            if(hasProjectile)cir.setReturnValue(true);
+            if(hasProjectile) return Optional.of(true);
         }
+
+        return Optional.empty();
     }
 }
